+++ src/action.ts
@@ -1,5 +1,1 @@
-export interface Action<T = any> {
-  type: any;
-  payload?: T;
-  [key: string]: any;
-}
+export class Action {}