+++ src/actions.ts
@@ -1,21 +1,17 @@
 import { BehaviorSubject } from "rxjs";
-import { filter } from "rxjs/operators";
+import { filter, map } from "rxjs/operators";
 import { Action } from "./action";
 
 export class Actions {
   constructor(private _dispatcher: BehaviorSubject<Action>) {}
 
-  whereType(actionType: any) {
+  isA<T extends Action>(actionOf: new () => T) {
     return this._dispatcher.pipe(
-      filter((action) => action.type === actionType)
+      filter((action) => action instanceof actionOf),
+      map((action) => action as T)
     );
   }
 
-  whereTypes(...actionTypes: any[]) {
-    return this._dispatcher.pipe(
-      filter((action) => actionTypes.includes(action.type))
-    );
-  }
   where(predicate: (action: Action) => boolean) {
     return this._dispatcher.pipe(filter(predicate));
   }