+++ src/index.ts
@@ -1,15 +1,4 @@
 import { createStore, RegisterState, MonoStore } from "./store";
 import { Actions } from "./actions";
 import { Action } from "./action";
-import { StateController } from "./stateController";
-import { Get, RemoveController } from "./provider";
-export {
-  createStore,
-  RegisterState,
-  Action,
-  MonoStore,
-  Actions,
-  StateController,
-  Get,
-  RemoveController,
-};
+export { createStore, RegisterState, Action, MonoStore, Actions };