+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 0.4.0
+
+```ts
 import { map, tap, delay } from "rxjs/operators";
-import { RegisterState, Action } from "../src";
+import { RegisterState, Action } from "mono-state";
 
 export interface Counter {
   count: number;
@@ -42,3 +45,38 @@ export class AsyncInc extends Action {}
 export const increment = () => new Inctrment();
 export const decrement = () => new Decrement();
 export const asyncInc = () => new AsyncInc();
+```
+
+## 0.3.0
+
+```ts
+import { RegisterState } from "mono-state";
+
+export interface Counter {
+  count: number;
+  loading: boolean;
+}
+export const counterState: RegisterState<Counter> = {
+  stateName: "counter",
+  initialState: { loading: false, count: 0 },
+  mapActionToState(emit) {
+    return {
+      inc(state) {
+        emit({ loading: false, count: state.count + 1 });
+      },
+      dec(state) {
+        emit({ loading: false, count: state.count - 1 });
+      },
+      async asyncInc(state) {
+        emit({ loading: true, count: state.count });
+        await delay(10);
+        emit((c_state) => ({ loading: false, count: c_state.count + 1 }));
+      },
+    };
+  },
+};
+
+function delay(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+```